# api/openapi.yaml
openapi: 3.0.3
info:
  title: Injury Intelligence API
  description: API for legal case assessment and attorney matching
  version: 1.0.0
  contact:
    name: Injury Intelligence API Support
    email: support@injuryintelligence.com
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.injuryintelligence.com
    description: Production server
paths:
...
  /v1/predict:
    post:
      summary: Predict case viability and value
      tags:
        - Predictions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assessmentId:
                  type: string
      responses:
        '200':
          description: Prediction results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResult'
  /v1/predict/simulate:
    post:
      summary: Simulate scenario changes
      tags:
        - Predictions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                base:
                  type: object
                toggles:
                  type: object
      responses:
        '200':
          description: Simulation results
          content:
            application/json:
              schema:
                type: object
                properties:
                  deltas:
                    type: object
...
components:
  schemas:
...
    Prediction:
      type: object
      properties:
        id:
          type: string
        model_version:
          type: string
        viability:
          type: object
          properties:
            overall:
              type: number
            liability:
              type: number
            causation:
              type: number
            damages:
              type: number
            ci:
              type: array
              items:
                type: number
        value_bands:
          type: object
          properties:
            p25:
              type: number
            median:
              type: number
            p75:
              type: number
        explainability:
          type: array
          items:
            type: object
            properties:
              feature:
                type: string
              direction:
                type: string
                enum: [+, -]
              impact:
                type: number
        created_at:
          type: string
          format: date-time
    PredictionResult:
      allOf:
        - $ref: '#/components/schemas/Prediction'
        - type: object
          properties:
            assessment_id:
              type: string
            model_version:
              type: string
            caveats:
              type: array
              items:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []